requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
piper-tts==1.3.0
python-multipart==0.0.6